// src/app/shared/enums/incident.type.ts
export enum IncidentStatus {
  INCIDENT_STATUS_UNSPECIFIED = 0,
  INCIDENT_STATUS_OPENED = 1,
  INCIDENT_STATUS_STUDIED = 2,
  INCIDENT_STATUS_CAN_BE_CLOSED = 3,
  INCIDENT_STATUS_CLOSED = 4,
}

export enum IncidentListSortBy {
  SORT_INCIDENT_LIST_UNSPECIFIED = 0,
  INCIDENT_LIST_BY_START_TIME = 1,
  INCIDENT_LIST_BY_END_TIME = 2,
}